add check for problem files for philter; fixed physionet tmp path
The physio net tmp path didn't include the short corpus details
so all corpora were getting mixed together.  Not a problem for smaller
batches and test runs but not technically a clean environment.

Philter runs forever on a handful of files.  We need to explore why
they are so problematic but for now, we'll just skip those
files during runs.