initial README; added date norm dirs; fixed NeuroNER dir; new *_BINs
Initial README covers the environment variables used in the the
shell script that should be set by the end user. This includes two
variables new to the shell script for specifying the Python bin
with the appropriate addtional packages installed for
ETUDE and PHILTER.

NeuroNER needed a patch in how it was processed to make sure we
always pick out the most recent run output directory.

NLM Scrubber now processes original corpora and date normalized
corpora (where years are randomly fixed to between 1950 and 2021.